dennisivy/requirements.txt
asgiref==3.2.3
boto3==1.14.62
botocore==1.17.62
certifi==2020.6.20
Django==3.0.3
django-filter==2.3.0
django-storages==1.10.1
docutils==0.15.2
Faker==4.1.1
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
smtp==0.1.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
whitenoise==5.2.0
wincertstore==0.2
xlwt==1.3.0
